requirements.txt
celery==3.1.26.post2
cycler==0.10.0
Django==3.0.6
django-celery==3.3.1
importlib-metadata==1.6.0
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
